// src/enums/enums.ts
export const enum PizzaSize {
  S = 'Small',
  M = 'Medium',
  L = 'Large',
  X = 'X-large',
}

export const enum PizzaMeatToppings {
  Chicken = 'Chicken',
  Hamburger = 'Hamburger',
  Peperoni = 'Peperoni',
  Sausage = 'Sausage',
}

export const enum PizzaVeggieToppings {
  BlackOlives = 'Black olives',
  GreenOlives = 'Green onions',
  Jalapenos = 'Jalepenos',
  Kimchi = 'Kimchi',
  Mushrooms = 'Mushrooms',
  Onions = 'Onions',
}

export const enum Toppings {
  Chicken = 'Chicken',
  Hamburger = 'Hamburger',
  Peperoni = 'Peperoni',
  Sausage = 'Sausage',
  BlackOlives = 'Black olives',
  GreenOlives = 'Green onions',
  Jalapenos = 'Jalepenos',
  Kimchi = 'Kimchi',
  Mushrooms = 'Mushrooms',
  Onions = 'Onions',
}

export const enum PizzaSauce {
  BBQ = 'bbq',
  Red = 'red',
  Oil = 'oil',
  White = 'white',
}